Alterations for solving the comments/suggestions on the pull request
