+++ aiounifi/interfaces/traffic_rules.py
@@ -33,5 +33,5 @@ class TrafficRules(APIHandler[TrafficRule]):
         traffic_rule_response = await self.controller.request(
             TrafficRuleEnableRequest.create(traffic_rule_dict, enable=state)
         )
-        self.controller.traffic_rules.process_typed_api_response(traffic_rule_response)
+        self.process_raw(traffic_rule_response.get("data", []))
         return traffic_rule_response
